Add simple parser.
Add graphite_parse() which takes either a line of input in the traditional
form "<metric> <value> <timestamp>" or just "<value> <timestamp>" and an
additional parameter for the metric, (for the case of AMQP where the metric
is the basically the same as a routing key and so may be separate).

If passed, the final parameter should be a three element array that is
populated with the pointers to the beginning of each part in the order
metric, value, timestamp.